backend/requirements.txt
# Web Framework
Flask==2.0.1
Flask-CORS==4.0.0
Flask-SocketIO==5.3.0
Flask-SQLAlchemy==2.5.1
werkzeug==2.0.3

# Database
SQLAlchemy==1.4.48
alembic==1.12.0
psycopg2-binary==2.9.0

# Async & Task Queue
celery==5.3.0
redis==5.0.0

# Validation & Serialization
marshmallow==3.20.0
marshmallow-sqlalchemy==0.29.0

# LLM Integration - LangChain/LangGraph (最新版)
langchain==1.2.10
langchain-core==1.2.12
langchain-openai==1.1.9
langchain-anthropic==1.3.3
langchain-google-genai==4.2.0
langchain-ibm==1.0.4
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-checkpoint-sqlite==3.0.3

# LLM Provider SDKs
openai==1.3.0
anthropic==0.7.0
ibm-watsonx-ai==1.4.4
google-generativeai==0.3.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
cryptography==41.0.0

# Development
pytest==7.4.0
pytest-cov==4.1.0
black==23.10.0
flake8==6.1.0
isort==5.12.0
mypy==1.6.0

# API Documentation
flask-restx==1.2.0